Add battery voltage measurement and improve GPS interaction
Don't read the whole binary sentence in a single tick. This takes
too long (~6ms). Read it byte by byte, this is still fast enough.

Also use two different structs to hold the data. The unpacked one
minimizes the number of expensive accesses to the packed struct.

Read battery voltage every 53th tick and don't do I2C communication
in the same tick as this tends to interfere with the battery voltage
measurement.